Ajustando implementação de cliente e reservas

+++ src/main/java/com/reservas/api/service/LeasesService.java
@@ -8,6 +8,7 @@ import com.reservas.api.entities.enums.ReservationStatus;
 import com.reservas.api.entities.model.Reservations;
 import com.reservas.api.entities.model.User;
 import com.reservas.api.exception.BusinessException;
+import com.reservas.api.exception.ForbiddenException;
 import com.reservas.api.exception.ResourceNotFoundException;
 import com.reservas.api.entities.mapper.LeasesMapper;
 import com.reservas.api.entities.model.Leases;
@@ -15,6 +16,8 @@ import com.reservas.api.repository.LeasesRepository;
 import com.reservas.api.repository.ReservationRepository;
 import com.reservas.api.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -33,8 +36,6 @@ public class LeasesService {
 	private final LeasesRepository leasesRepository;
 	private final LeasesMapper leasesMapper;
 	private final ReservationRepository reservationRepository;
-	private final UserRepository userRepository;
-	private final ReservationMapper reservationMapper;
 
 	@Transactional(readOnly = true)
 	public List<LeasesResponse> listAll() {
@@ -58,34 +59,43 @@ public class LeasesService {
 		});
 
 		if (leasesRequest.getMinTime() > leasesRequest.getMaxTime()) {
-			throw new BusinessException("Leases minimum time is " + leasesRequest.getMinTime());
+			throw new BusinessException("Min time must be less than max time");
 		}
 
-		Leases newLeases = leasesMapper.toEntity(leasesRequest);
-		newLeases.setCreatedAt(LocalDate.now());
+		User owner = getCurrentAuthenticatedUser();
 
-		Leases savedLeases = leasesRepository.save(newLeases);
-		return leasesMapper.toResponse(savedLeases);
-	}
+		Leases newLease = leasesMapper.toEntity(leasesRequest);
+		newLease.setOwner(owner);
+
+		Leases savedLease = leasesRepository.save(newLease);
+		return leasesMapper.toResponse(savedLease);
+ 	}
 
 	@Transactional
 	public LeasesResponse update(UUID id, LeasesRequest leasesRequest) {
-		 Leases existingLeases = leasesRepository.findById(id)
-				 .orElseThrow(() -> new ResourceNotFoundException("Leases type not found with id: " + id));
+		Leases existingLease = leasesRepository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("Leases type not found with id: " + id));
 
-		 leasesRepository.findByName(leasesRequest.getName()).ifPresent(lease -> {
-			 if (!lease.getId().equals(id)) {
-				 throw new BusinessException("Leases with name " + leasesRequest.getName() + " already exists");
-			 }
-		 });
+		checkOwnershipOrAdmin(existingLease.getOwner());
 
-		 if (leasesRequest.getMinTime() > leasesRequest.getMaxTime()) {
-			 throw new BusinessException("Min time must be less than max time");
-		 }
+		leasesRepository.findByName(leasesRequest.getName()).ifPresent(lease -> {
+			if (!lease.getId().equals(id)) {
+				throw new BusinessException("Leases with name " + leasesRequest.getName() + " already exists");
+			}
+		});
 
-		 leasesMapper.updateEntityFromRequest(leasesRequest, existingLeases);
-		 Leases leasesUpdated = leasesRepository.save(existingLeases);
-		 return leasesMapper.toResponse(leasesUpdated);
+		if (leasesRequest.getMinTime() != null && leasesRequest.getMaxTime() != null &&
+				leasesRequest.getMinTime() > leasesRequest.getMaxTime()) {
+			throw new BusinessException("Minimum time cannot be greater than maximum time.");
+		} else if(leasesRequest.getMinTime() != null && leasesRequest.getMinTime() > existingLease.getMaxTime()) {
+			throw new BusinessException("Minimum time cannot be greater than maximum time.");
+		} else if (leasesRequest.getMaxTime() != null && existingLease.getMinTime() > leasesRequest.getMaxTime()) {
+			throw new BusinessException("Minimum time cannot be greater than maximum time.");
+		}
+
+		leasesMapper.updateEntityFromRequest(leasesRequest, existingLease);
+		Leases updatedLease = leasesRepository.save(existingLease);
+		return leasesMapper.toResponse(updatedLease);
 	}
 
 	@Transactional
@@ -93,8 +103,10 @@ public class LeasesService {
 		Leases leases = leasesRepository.findById(id)
 				.orElseThrow(() -> new ResourceNotFoundException("Leases type not found with id: " + id));
 
+		checkOwnershipOrAdmin(leases.getOwner());
+
 		if (reservationRepository.existsByLeases(leases)) {
-			throw new BusinessException("Cannot be deleted. This rental type already has associated reservations");
+			throw new BusinessException("Leases cannot be deleted as it has associated reservations");
 		}
 
 		leasesRepository.delete(leases);
@@ -126,55 +138,21 @@ public class LeasesService {
 		                        .collect(Collectors.toList());
 	}
 
-	@Transactional
-	public ReservationResponse hireLease(UUID leaseId, UUID userId, LocalDateTime startDate, LocalDateTime endDate) {
-
-		if (startDate == null || endDate == null) {
-			throw  new BusinessException("Start date and end date cannot be null");
+	private User getCurrentAuthenticatedUser() {
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		if (authentication == null || !authentication.isAuthenticated() || !(authentication.getPrincipal() instanceof User)) {
+			throw new IllegalStateException("User not authenticated correctly.");
 		}
+		return (User) authentication.getPrincipal();
+	}
 
-		if (endDate.isBefore(startDate) || endDate.isEqual(startDate)) {
-			throw new BusinessException("End date must be after start date");
+	private void checkOwnershipOrAdmin(User owner) {
+		User currentUser = getCurrentAuthenticatedUser();
+		boolean isAdmin = currentUser.getAuthorities().stream()
+		                             .anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN"));
+
+		if (!isAdmin && !currentUser.getId().equals(owner.getId())) {
+			throw new ForbiddenException("Access denied. You can only manage your own locacoes.");
 		}
-
-		if (startDate.toLocalDate().isBefore(LocalDate.now())) {
-			throw new BusinessException("Start date cannot be in the past");
-		}
-
-		Leases leases = leasesRepository.findById(leaseId)
-				.orElseThrow(() -> new ResourceNotFoundException("Leases type not found with id: " + leaseId));
-
-		User user = userRepository.findById(userId)
-				.orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + userId));
-
-		boolean hasConflictingReservation = reservationRepository
-				.existsByLeasesAndStatusInAndStartDateLessThanEqualAndEndDateGreaterThanEqual(
-						leases,
-						List.of(ReservationStatus.PENDING, ReservationStatus.CONFIRMED),
-						endDate,
-						startDate
-		);
-
-		if (hasConflictingReservation) {
-			throw  new BusinessException("This property is already reserved for the selected period");
-		}
-
-		long numberOfDays = ChronoUnit.DAYS.between(startDate.toLocalDate(), endDate.toLocalDate()) + 1;
-		BigDecimal dailyPrice = leases.getHourValue();
-		BigDecimal totalValue = dailyPrice.multiply(BigDecimal.valueOf(numberOfDays));
-
-		Reservations reservation = new Reservations();
-		reservation.setLeases(leases);
-		reservation.setUser(user);
-		reservation.setStartDate(startDate);
-		reservation.setEndDate(endDate);
-		reservation.setStatus(ReservationStatus.PENDING);
-		reservation.setCreatedAt(LocalDateTime.now());
-		reservation.setUpdatedAt(LocalDateTime.now());
-		reservation.setTotalValue(totalValue);
-
-		Reservations savedReservation = reservationRepository.save(reservation);
-
-		return reservationMapper.toResponse(savedReservation);
 	}
 }
